Script to clean up OSISM Testbed heat stack.
OTC's heat ("RTS") is outdated and broken.
The cascading setup of OTC brings several resource dependencies that
heat would need to observe when cleaning up resources. However, heat
has not been adjusted, so it miserably fails. This helper script
allows to still clean up by helping heat a little ...
It has hardcoded the server and security group names from the OSISM
testbed installation, so it's specialized for that use case.

Signed-off-by: Kurt Garloff <scs@garloff.de>