Add OTC internal nameserver to management subnet.
This allows an internal connection to the OpenStack APIs on OTC.
We should have better connectivity and performance from internal DNS
a well as a direct connection to the API gateway. (On OTC, all OpenStack
API endpoints are routed via one central gateway.)
This hardcodes the OTC DNS server address and is thus not generic.
However, we might need a generic mechanism as potentially some other
clouds will benefit from (or even have a need to) using an internal DNS
server.

Signed-off-by: Kurt Garloff <scs@garloff.de>