Rename vNICs from enp4sX to ensX.
The secondary virtual network cards are attached to a secondary PCI bus
which results in Ubu 18.04 naming them enp4sX rather than ensX
(and not bringing them up by default).
This breaks many assumptions in OSISM testbed, as the NIC names are
hardcoded at a number of places. So rename the vNICs early on and
bring them up.

The changes will be a no-op on most clouds, so they could be merged.

A more generic mechanism would be better, as the topology and
consequently naming of vNICs will very between clouds. Ideally, we
could also autodetect things, but it's not trivial I fear.

Signed-off-by: Kurt Garloff <scs@garloff.de>