Pass --timeout when creating stacks.
We run into the default timeout of 3600(s) when deploying
infra+ceph+openstack. So rather pass an explicit timeout,
to avoid hitting it. 50mins for just the deployment, 80mins
for infra, 110mins for infra+ceph and 150mins for
infra+ceph+openstack are chosen such, that we should never
see timeouts on the top level. (The wait conditions with
1800s and 7200s should hit first, though they are conservatively
chosen as well.)

Signed-off-by: Kurt Garloff <scs@garloff.de>