Add the network interface name hacks.
This is just a literal copy from the hacks that are needed to name the
NICs in OTC as the scripts expect it.  Unlike most of the other changes
to get OTC to work, this is NOT a strangeness in OTC, but just a lack of
flexibility / autodetection in OSISM testbed.